refactor: remove unused code, reorder
