+++ package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "start": "node src/index.js -- --username=malinka"
+    "start": "node src/index.js --"
   },
   "author": "malinka775",
   "license": "ISC",
