+++ src/index.js
@@ -1,1 +1,7 @@
-import './start/index.js'
+import { start } from './start/index.js'
+
+try {
+  await start();
+} catch (e) {
+  console.log(e.message);
+}
